+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Client
     - Added hover trigger mode for `Popup` view which enables with options to constructor, e.g. `{ hoverTriggers: string, hoverElementToOptions: fn(triggerEl) }`
+    - Fixed `Popup` closing on click on/inside a trigger element while a popup is visible, now it stays visible
     - Fixed `Popup` z-index when several popups are shown
     - Fixed `Popup` to set `max-width` for its content
     - Added `signature` view